feat(mission5): 게시글 수정 시 기존 게시글의 제목과 본문 내용을 가져와서 출력하는 기능 구현

+++ src/main/java/kr/codesqaud/cafe/repository/ArticleRepository.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 
 public interface ArticleRepository {
     void saveArticle(Article article);
-    Optional<Article> findOneArticleById(int id);
+    Optional<Article> findOneArticleById(long id);
     List<Article> getArticles();
     void deleteArticle(Member member);
     int getSize();
